Windows port: commit updated files and add util_unix.go/util_windows.go
